+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
+## 1.2.2
+- Fixed a bug when decoding negative integers ([#73](https://github.com/simolus3/web3dart/issues/73))
+
 ## 1.2.0
 - Added `estimateGas` method on `Web3Client` to estimate the amount of gas that
   would be used by a transaction.
   
-In 1.2.1, the `atBlock` parameter on `estimateGas` was deprecated an will be ignored.
+In 1.2.1, the `atBlock` parameter on `estimateGas` was deprecated and will be ignored.
 
 ## 1.1.1, 1.1.1+1
 - Fix parsing transaction receipts when the block number is not yet available.